extracts network stack into its own library
