+++ app/build.gradle
@@ -45,6 +45,7 @@ android {
 
 dependencies {
     implementation fileTree(dir: "libs", include: ["*.jar"])
+    implementation project(':network')
 
     kapt Dependencies.roomCompiler
     kapt Dependencies.glideCompiler
