+++ app/src/main/java/com/example/architecture/android/core/datasource/TypiCodeDataSource.kt
@@ -1,6 +1,7 @@
-package com.example.architecture.android.core.network
+package com.example.architecture.android.core.datasource
 
-import com.example.architecture.android.core.data.BaseDataSource
+import com.example.architecture.android.core.base.BaseDataSource
+import com.example.network.TypiCodeService
 import javax.inject.Inject
 
 class TypiCodeDataSource @Inject constructor(
