+++ app/src/main/java/com/example/architecture/android/core/di/AppComponent.kt
@@ -4,6 +4,8 @@ import android.app.Application
 import com.example.architecture.android.DaggerSampleApplication
 import com.example.architecture.android.core.di.modules.AppModule
 import com.example.architecture.android.core.di.modules.CoroutinesModule
+import com.example.network.di.NetworkModule
+import com.example.network.di.WebSocketModule
 import dagger.BindsInstance
 import dagger.Component
 import dagger.android.AndroidInjectionModule
@@ -14,7 +16,9 @@ import javax.inject.Singleton
     modules = [
         AndroidInjectionModule::class,
         AppModule::class,
-        CoroutinesModule::class
+        CoroutinesModule::class,
+        NetworkModule::class,
+        WebSocketModule::class
     ]
 )
 interface AppComponent {
