+++ app/src/main/java/com/example/architecture/android/landing/LandingRepository.kt
@@ -1,8 +1,8 @@
 package com.example.architecture.android.landing
 
 import com.example.architecture.android.core.data.Result
-import com.example.architecture.android.core.network.TypiCodeDataSource
-import com.example.architecture.android.core.network.model.response.Post
+import com.example.architecture.android.core.datasource.TypiCodeDataSource
+import com.example.network.model.response.Post
 import javax.inject.Inject
 
 interface ILandingRepository {
