+++ build.gradle
@@ -2,6 +2,9 @@ import com.example.buildsrc.Versions
 
 buildscript {
 
+    ext {
+        kotlin_version = '1.3.61'
+    }
     repositories {
         google()
         jcenter()
