+++ network/src/main/java/com/example/network/di/NetworkModule.kt
@@ -1,0 +1,65 @@
+package com.example.network.di
+
+import com.example.network.TypiCodeService
+import com.example.network.websocket.WebSocketService
+import com.google.gson.Gson
+import com.tinder.scarlet.Scarlet
+import dagger.Module
+import dagger.Provides
+import okhttp3.OkHttpClient
+import okhttp3.logging.HttpLoggingInterceptor
+import retrofit2.Retrofit
+import retrofit2.converter.gson.GsonConverterFactory
+import javax.inject.Singleton
+
+@Module
+class NetworkModule {
+
+    @Provides
+    fun provideOkHttpClient(interceptor: HttpLoggingInterceptor): OkHttpClient =
+        OkHttpClient.Builder().addInterceptor(interceptor)
+            .build()
+
+    @Provides
+    fun provideLoggingInterceptor() = HttpLoggingInterceptor().apply {
+        HttpLoggingInterceptor.Level.BODY
+    }
+
+    @Provides
+    @Singleton
+    fun provideGson(): Gson = Gson()
+
+    @Provides
+    @Singleton
+    fun provideGsonConverterFactory(gson: Gson): GsonConverterFactory =
+        GsonConverterFactory.create(gson)
+
+    @Singleton
+    @Provides
+    fun providesTypiCodeService(
+        okhttpClient: OkHttpClient,
+        converterFactory: GsonConverterFactory
+    ) = provideService(okhttpClient, converterFactory, TypiCodeService::class.java)
+
+    @Singleton
+    @Provides
+    fun providesWebSocketService(scarlet: Scarlet) = scarlet.create(WebSocketService::class.java)
+
+    private fun createRetrofit(
+        okhttpClient: OkHttpClient,
+        converterFactory: GsonConverterFactory
+    ): Retrofit {
+        return Retrofit.Builder()
+            .baseUrl(TypiCodeService.BASE_URL)
+            .client(okhttpClient)
+            .addConverterFactory(converterFactory)
+            .build()
+    }
+
+    private fun <T> provideService(
+        okhttpClient: OkHttpClient,
+        converterFactory: GsonConverterFactory, clazz: Class<T>
+    ): T {
+        return createRetrofit(okhttpClient, converterFactory).create(clazz)
+    }
+}
