+++ settings.gradle
@@ -1,1 +1,2 @@
+include ':network'
 include ':app'
